misc: moved google eye logic to different folder
